Remove program error footgun and cleaner developer experience (#8042)

+++ programs/bpf/c/src/panic/panic.c
@@ -4,7 +4,7 @@
  */
 #include <solana_sdk.h>
 
-extern uint32_t entrypoint(const uint8_t *input) {
+extern uint64_t entrypoint(const uint8_t *input) {
   sol_panic();
   return SUCCESS;
 }
