+++ programs/bpf/c/src/relative_call/relative_call.c
@@ -8,7 +8,7 @@ void __attribute__ ((noinline)) helper() {
   sol_log(__func__);
 }
 
-extern uint32_t entrypoint(const uint8_t *input) {
+extern uint64_t entrypoint(const uint8_t *input) {
   sol_log(__func__);
   helper();
   return SUCCESS;
