+++ programs/bpf/rust/128bit/src/lib.rs
@@ -4,7 +4,7 @@ extern crate solana_sdk;
 use solana_sdk::entrypoint::SUCCESS;
 
 #[no_mangle]
-pub extern "C" fn entrypoint(_input: *mut u8) -> u32 {
+pub extern "C" fn entrypoint(_input: *mut u8) -> u64 {
     let x: u128 = 1;
     let y = x.rotate_right(1);
     assert_eq!(y, 170_141_183_460_469_231_731_687_303_715_884_105_728);
