+++ programs/bpf/rust/alloc/src/lib.rs
@@ -7,7 +7,7 @@ use solana_sdk::{entrypoint::SUCCESS, info};
 use std::{alloc::Layout, mem};
 
 #[no_mangle]
-pub extern "C" fn entrypoint(_input: *mut u8) -> u32 {
+pub extern "C" fn entrypoint(_input: *mut u8) -> u64 {
     unsafe {
         // Confirm large allocation fails
