+++ programs/bpf/rust/iter/src/lib.rs
@@ -4,7 +4,7 @@ extern crate solana_sdk;
 use solana_sdk::{entrypoint::SUCCESS, info};
 
 #[no_mangle]
-pub extern "C" fn entrypoint(_input: *mut u8) -> u32 {
+pub extern "C" fn entrypoint(_input: *mut u8) -> u64 {
     const ITERS: usize = 100;
     let ones = [1_u64; ITERS];
     let mut sum: u64 = 0;
