+++ programs/bpf/rust/param_passing/src/lib.rs
@@ -5,7 +5,7 @@ use solana_bpf_rust_param_passing_dep::{Data, TestDep};
 use solana_sdk::{entrypoint::SUCCESS, info};
 
 #[no_mangle]
-pub extern "C" fn entrypoint(_input: *mut u8) -> u32 {
+pub extern "C" fn entrypoint(_input: *mut u8) -> u64 {
     let array = [0xA, 0xB, 0xC, 0xD, 0xE, 0xF];
     let data = Data {
         twentyone: 21u64,
